app/requirements.txt
appdirs
asgiref
black
certifi
cffi
chardet
click
coverage
crispy-bootstrap5
cryptography
defusedxml
Django
django-allauth
django-cors-headers
django-crispy-forms
django-email-verification
django-environ
djangorestframework
idna
mypy-extensions
oauthlib
pathspec
Pillow
psycopg2-binary
pycparser
PyJWT
python3-openid
pytz
regex
requests
requests-oauthlib
sqlparse
toml
urllib3